Replaced RPC Ai check with Check Transactions
