+++ runtime/Cargo.toml
@@ -42,6 +42,7 @@ rand = { workspace = true }
 rayon = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true, features = ["rc"] }
+serde_json = { workspace = true }
 serde_derive = { workspace = true }
 domichain-address-lookup-table-program = { workspace = true }
 domichain-bpf-loader-program = { workspace = true }
